Update VideoPlayer2.tsx
Tracks the current video index using the currentVideoIndex state variable.
Accepts a list of videos as input (videoList) and uses currentVideoIndex to determine the current video.
When a video ends, it checks whether there is a next video. If there is, it automatically plays the next video using handleVideoEnded.
Adds a button that appears when there is a next video in the list. When clicked, it advances to the next video in the list.