FEAT: Add calendar, Fixed Lint errors
In this i have added a calendar component for the Cohort students and admin can access and update events, so that students can just come and check on which day which topic in class happened and find its video link and Notes attached. And Students can see upcoming agenda at the single place that is updated by the admin of the next entire 30 days from the current date.