`docs: add note about using --depth=1 for cloning issues`
This update improves the documentation by adding a note on using the `--depth=1` flag when cloning the repository. Due to large repo size, simple clone may not work for many devs. The note has been added in both the "Running Locally" and "Contributing" sections to ensure clarity for all contributors.